refactor(ui-react): split session list modules
