+++ apps/desktop/src/main/__tests__/radius-converge-contract.test.ts
@@ -186,9 +186,9 @@ const COMPONENT_RADIUS: ComponentRadiusCheck[] = [
   { file: 'packages/ui/src/primitives/menu.tsx', name: 'MenuPopup', tier: 'surface' },
   { file: 'packages/ui/src/primitives/alert.tsx', name: 'alertVariants', tier: 'surface' },
   { file: 'packages/ui/src/primitives/toolbar.tsx', name: 'Toolbar', tier: 'surface' },
-  { file: 'packages/ui/src/session-list-panel.tsx', name: 'navRowVariants', tier: 'control' },
-  { file: 'packages/ui/src/session-list-panel.tsx', name: 'settingsButtonClass', tier: 'control' },
-  { file: 'packages/ui/src/session-list-panel.tsx', name: 'rowActionVariants', tier: 'control' },
+  { file: 'packages/ui/src/session-sidebar-nav.tsx', name: 'navRowVariants', tier: 'control' },
+  { file: 'packages/ui/src/session-sidebar-nav.tsx', name: 'settingsButtonClass', tier: 'control' },
+  { file: 'packages/ui/src/session-history-list.tsx', name: 'rowActionVariants', tier: 'control' },
 ];
 
 /** Extract the body of a component declaration by brace-matching from the
